- Added Span<byte> and ReadonlySequence<byte> overrides with encoding - Fixed MutliByte (cchar_t) writes with an advance method to get consistent behaviour
git-svn-id: file:///H/Users/C_B-A/temp/NCurses.Core_git/branches/span@71 3a9065fc-4a7e-7541-9dc8-748989605d84

+++ src/NCurses.Core/Interop/Dynamic/cchar_t/cchar_tBuilder.cs
@@ -708,7 +708,7 @@ namespace NCurses.Core.Interop.Dynamic.cchar_t
 
             methodIl.Emit(OpCodes.Nop);
             methodIl.Emit(OpCodes.Ldarg_0);
-            methodIl.Emit(OpCodes.Ldarga_S, 1);
+            methodIl.Emit(OpCodes.Ldarga_S, (byte)1);
             methodIl.Emit(OpCodes.Call, opEquality);
             methodIl.Emit(OpCodes.Stloc_0);
             methodIl.Emit(OpCodes.Br_S, lbl1);
